requirements.txt
flask
pillow
pytesseract
beautifulsoup4
lxml
requests
serpapi
